requirements.txt
numpy
scipy
pyquaternion
opencv-python
h5py
psutil
tqdm
termcolor
tensorboard
tensorboardX
wandb
gymnasium
dlimp@git+https://github.com/ojh6404/dlimp.git
imageio
imageio-ffmpeg
matplotlib
jax
flax
optax
orbax
distrax
chex
clu
absl-py
Pillow
ml-collections
tensorflow-datasets
# tensorflow[and-cuda]
tensorflow_hub
tensorflow_text
huggingface_hub
transformers
einops
omegaconf
moviepy
plotly
octo@git+https://github.com/ojh6404/octo.git
